requirements.txt
langchain
langchain-openai
langchain-community
pydantic
requests
python-dotenv
langgraph
duckduckgo-search
streamlit